refactor(routes): extract /admin into admin.js (god-file decomposition)
18 admin routes moved verbatim to server/routes/admin.js, wired via registerAdminRoutes.
Route contract unchanged (284 routes), 0 /admin left inline, vitest green. Lucy (GLM-5.2), byte-reviewed by m5Max.
